refactor(roi): simplify-pass cleanups on Part B upscale code
Fixes from /simplify review of B.1-B.3:

- Align RescueHammingTooLax error variant field types (both u32) so the
  rescue and primary Hamming budgets render consistently.
- Elide per-pixel bounds checks in the Lanczos horizontal pass by
  pre-slicing `src_row[bx..bx+bw]` once per ROI row.
- Drop narrating "Horizontal pass / Vertical pass" WHAT-comments; keep
  the clamping WHY-comment that justifies the tap-index saturation.
- Refactor the 6-tap accumulators to `.iter().zip()` and remove the
  `clippy::needless_range_loop` allow.

+++ crates/locus-core/src/config.rs
@@ -407,7 +407,7 @@ impl DetectorConfig {
             }
             if u32::from(r.rescue_max_hamming) >= self.max_hamming_error {
                 return Err(ConfigError::RescueHammingTooLax {
-                    rescue: r.rescue_max_hamming,
+                    rescue: u32::from(r.rescue_max_hamming),
                     primary: self.max_hamming_error,
                 });
             }
